+++ CHANGELOG.md
@@ -2,6 +2,7 @@ English | [简体中文](./CHANGELOG_CN.md)
 
 #### V3.4.1 (2021-xx-xx)
 
+- [FEATURE] Add `setSwitchPosition(x, y)` method to update the position of switch button, see [Tutorial](./doc/public_properties_methods.md) for more details.
 - [FIX] Update theme style after `setOption()`.
 - [FIX] Add `Symbol` polyfill. (issue #361)
 - [FIX] Fix `fetch` optional parameter `init`. (issue #363, #365)